Added Redux/React-Redux, Added a NewLocation modal and linked them together
Added redux and two stores: one of just a boolean to display the AddLocation modal and a second to store all the locations the user can swipe through.

Pressing the + button triggers the modal store to toggle and show the window. Adding a location from the modal will update the locationstore with a new location.

TODO:
* Add removing from the store to the UI.
* Link the locations to to left and wipe swiping