Swapped ViewPagerAndroid to FlatList to work on all platforms. Various UI platorm fixes
Stopped using the ViewPagerAndroid because it only works on as single platform. Instead swapping to a FlatList. Should work on Android, ios, UWP and Mac.

Swapped Location string to a MappedLocation object that contains everything a location needs. Not just a zipcode.

Few minor UI rendering changes. Added a button to AddNewLocation to exit. Fixed the weird half circle rendering CSS differences between IOS and android.